Bug 1674994 part 1 - Make Atomics work with large ArrayBuffers. r=lth
The remaining three uint32_t uses in AtomicsObject.h/cpp are unrelated.

This also fixes the byteLength getter on ArrayBuffer and SharedArrayBuffer.

Differential Revision: https://phabricator.services.mozilla.com/D95688

UltraBlame original commit: 2db46505e947577170e24f2237af0cd93f14463a

+++ js/src/builtin/AtomicsObject.h
@@ -123,20 +123,20 @@ class FutexThread {
 
 
 MOZ_MUST_USE FutexThread::WaitResult atomics_wait_impl(
-    JSContext* cx, SharedArrayRawBuffer* sarb, uint32_t byteOffset,
-    int32_t value, const mozilla::Maybe<mozilla::TimeDuration>& timeout);
+    JSContext* cx, SharedArrayRawBuffer* sarb, size_t byteOffset, int32_t value,
+    const mozilla::Maybe<mozilla::TimeDuration>& timeout);
 
 
 MOZ_MUST_USE FutexThread::WaitResult atomics_wait_impl(
-    JSContext* cx, SharedArrayRawBuffer* sarb, uint32_t byteOffset,
-    int64_t value, const mozilla::Maybe<mozilla::TimeDuration>& timeout);
+    JSContext* cx, SharedArrayRawBuffer* sarb, size_t byteOffset, int64_t value,
+    const mozilla::Maybe<mozilla::TimeDuration>& timeout);
 
 
 
 
 
 MOZ_MUST_USE int64_t atomics_notify_impl(SharedArrayRawBuffer* sarb,
-                                         uint32_t byteOffset, int64_t count);
+                                         size_t byteOffset, int64_t count);
 
 }
